refactor: add header folder
