+++ src/components/Header/SocialButtons.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script setup lang="ts">
-import { useThemeStore } from "../store/theme";
+import { useThemeStore } from "../../store/theme";
 
 const themeStore = useThemeStore();
 </script>
