+++ src/components/Header/ToggleSelector.vue
@@ -7,7 +7,7 @@
 </template>
 
 <script setup lang="ts">
-import { useThemeStore } from "../store/theme";
+import { useThemeStore } from "../../store/theme";
 
 const themeStore = useThemeStore();
